Updated .properties and data type according to the database column type and login

+++ src/main/java/it/pc/test/WebSpringApp/dto/ProduttoreDTO.java
@@ -23,7 +23,7 @@ public class ProduttoreDTO extends AbstractBaseDTO<Integer> {
     private PartnerDTO partner;
 
     private TipoCategoria categoria;
-    private Double numeroDipendenti;
+    private Integer numeroDipendenti;
     private Double fatturatoAnnuo;
     private SedeDTO sede;
     private List<ProdottoDTO> prodotti;
